requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy
pyyaml
librosa
scipy
matplotlib
pytest
